Competing Risk Calibration

This function will allow you to compute calibration metrics: - predicted and observed probabilities - predicted and observed events of interest

calibrate(data, horizon, groups = 10, cause = 1, model)

Arguments

data
Dataframe to use for absolute risk predictions

horizon
Time horizon for predictions

groups
Number of deciles based on model-predicted risks (default is 10)

cause
Integer value for the event-of-interest to calibrate predictions

model
The returened model object from riskRegression

Value

A list containing three elements. The first element returns a string detailing the time horizon for calibration. The second element is a chi-sq test statistic and p-value for the Hosmer-Lemeshow test (with DF=groups-1). The third element is a tibble containining the decile index along with the predicted and observed cumulative incidences, and the predicted and observed events across the n risk quantiles.

Examples

library(riskRegression)

#> Loading required package: data.table

#> Loading required package: ggplot2

#> Loading required package: prodlim

#> riskRegression version 2018.10.03

# Simulate Data
sim.data <- sampleData(100, outcome = 'competing.risks')

# Fit Fine-Gray Model
fgr.mod <- FGR(formula = Hist(time, event) ~ X7 + X6 + X8 + X9,
               data = sim.data, cause = 1)

calibrate(sim.data, 2, 10, 1, fgr.mod)

#> Error: Column `pred` must be a 1d atomic vector or a list