Optimal ROC Threshold

Functions to identify the optimal ROC threshold according to various optimization criteria.

index_of_union(sens = NULL, spec = NULL, thr = NULL, auc = NULL)

youden(sens = NULL, spec = NULL, thr = NULL)

top_left(sens = NULL, spec = NULL, thr = NULL)

min_pval(time = NULL, status = NULL, pred = NULL)

conc_prob(sens = NULL, spec = NULL, thr = NULL)

Arguments

sens
Sensitivities

spec
Specificities

thr
Threshold values for ROC

auc
Area under the ROC curve

time
Time

status
Censoring status

pred
Predictions

Value

A scalar containing the chosen threshold.

Details

Youden's J Index: This function will compute Youden's J index. This is computed by finding the threshold which maximizes the value (sensitivity + specificity - 1)

Index of Union: This function will compute the Index of Union statistic which represents the ideal ROC threshold according to maximizing the statistic (sensitivity - auc) + (specificity - auc)

Top Left: Provides the receiver operating characteristic (ROC) threshold (x-axis) that minimizes the distance from the ROC curve to the to the top left corner of the plot (1,1)

Minimum P-value: Provides the receiver operating characteristic (ROC) threshold that minimizes the P-value according to Unal (2017), Article ID 3762651

Concordance Probability: Provides the receiver operating characteristic (ROC) threshold that maximizes the concordance probability. The concordance probability is the product of the sensitivity and specificity.

Examples

# Simulate some data
sens <- runif(100,0,1)
spec <- runif(100,0,1)
thr <- runif(100,0,1)
auc <- runif(1,0,1)

index_of_union(sens, spec, thr, auc)

#> Finding optimal threshold that minimizes the Index of Union
#>           function...

#> Search complete

#> [[1]]
#> [1] "Index of Union (Unal, 2017)"
#> 
#> $index.threshold
#> [1] 0.4002089
#> 

youden(sens, spec, thr)

#> Finding optimal threshold that maximizes Youden's J index...

#> 

#> 

#> Search complete

#> [[1]]
#> [1] "Youden's J Index"
#> 
#> $Youden.threshold
#> [1] 0.4002089
#> 

top_left(sens, spec, thr)

#> Finding optimal threshold that minimizes the Euclidean distance...

#> 

#> 

#> Search complete

#> [[1]]
#> [1] "Closest to top left"
#> 
#> $topleft.threshold
#> [1] 0.4002089
#> 

conc_prob(sens, spec, thr)

#> Finding optimal threshold that maximizes the Concordance
#>           Probability (product)...

#> 

#> 

#> Search complete

#> [[1]]
#> [1] "Concordance Probability Method (CZ)"
#> 
#> $cz.threshold
#> [1] 0.4002089
#>